requirements.txt
aiofiles==23.2.1
accelerate==0.27.2
bcrypt==4.1.2
bitsandbytes==0.41.3
diffusers==0.25.1
einops==0.7.0
fastapi==0.109.1
ftfy==6.1.3
huggingface-hub==0.20.3
imagesize==1.4.1
numpy<2.0.0
opencv-python-headless
passlib==1.7.4
pillow==10.2.0
prodigyopt
pydantic==2.6.0
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
rich==13.7.0
safetensors==0.4.2
scipy==1.11.4
tensorboard==2.15.1
timm==0.9.12
torch==2.1.2
torchvision==0.16.2
toml
tqdm==4.66.1
transformers==4.37.2
uvicorn==0.27.0
voluptuous==0.14.1
xformers==0.0.23.post1 